extract hub issue parsing to its own function
